refactor: ClipboardItemListコンポーネント機能分割リファクタリング
コンポーネント分割:
- ClipboardItem: 個別アイテム表示
- ClipboardContentRenderer: コンテンツ種別表示
- ClipboardStateViews: エラー・ローディング・空状態表示

カスタムフック作成:
- useImageWindow: 画像ウィンドウ表示ロジック
- useClipboardFormats: フォーマット管理ロジック

改善効果:
- メインコンポーネント374行→62行 (83%削減)
- 単一責任原則による保守性向上
- 再利用可能な設計
